[US-REFACTOR-RABEB] creation interfaces
