+++ excelservice-1.0/src/main/java/com/newlight77/kata/survey/controler/SurveyController.java
@@ -2,47 +2,55 @@ package com.newlight77.kata.survey.controler;
 
 import com.newlight77.kata.survey.model.Campaign;
 import com.newlight77.kata.survey.model.Survey;
-import com.newlight77.kata.survey.service.ExportCampaignService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.newlight77.kata.survey.service.impl.ExportCampaignServiceImpl;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 public class SurveyController {
 
-    private ExportCampaignService exportCampaignService;
+    private ExportCampaignServiceImpl exportCampaignService;
 
-    public SurveyController(final ExportCampaignService exportCampaignService) {
-      this.exportCampaignService = exportCampaignService;
+    private static final String uriCreateSurvey="/api/survey";
+
+    private static final String uriGetSurvey="/api/survey/{id}";
+
+    private static final String uriCreateCampaign="/api/survey/campaign";
+
+    private static final String uriGetCampaign="/api/survey/campaign/{id}";
+
+    private static final String uriExportCampaign="/api/survey/campaign/export";
+
+
+    public SurveyController(final ExportCampaignServiceImpl exportCampaignService) {
+        this.exportCampaignService = exportCampaignService;
     }
 
-    @RequestMapping(value = "/api/survey/create", method = RequestMethod.POST)
+    @RequestMapping(value = uriCreateSurvey,  method = RequestMethod.POST)
     public void createSurvey(@RequestBody Survey survey) {
         exportCampaignService.creerSurvey(survey);
     }
 
-    @RequestMapping(value = "/api/survey/get", method = RequestMethod.GET)
-    public Survey getSurvey(@RequestParam String id) {
+    @RequestMapping(value = uriGetSurvey, method = RequestMethod.GET)
+    public Survey getSurvey(@RequestParam  @PathVariable("id") String id) {
         return exportCampaignService.getSurvey(id);
     }
 
-    @RequestMapping(value = "/api/survey/campaign/create", method = RequestMethod.POST)
+    @RequestMapping(value = uriCreateCampaign, method = RequestMethod.POST)
     public void createCampaign(@RequestBody Campaign campaign) {
         exportCampaignService.createCampaign(campaign);
     }
 
-    @RequestMapping(value = "/api/survey/campaign/get", method = RequestMethod.GET)
-    public Campaign getCampaign(@RequestParam String id) {
+    @RequestMapping(value = uriGetCampaign, method = RequestMethod.GET)
+    public Campaign getCampaign(@RequestParam @PathVariable("id") String id) {
         return exportCampaignService.getCampaign(id);
     }
 
-    @RequestMapping(value = "/api/survey/campaign/export", method = RequestMethod.POST)
+    @RequestMapping(value = uriExportCampaign, method = RequestMethod.POST)
     public void exportCampaign(@RequestParam String campaignId) {
-
         Campaign campaign = exportCampaignService.getCampaign(campaignId);
         Survey survey = exportCampaignService.getSurvey(campaign.getSurveyId());
         exportCampaignService.sendResults(campaign, survey);
-        
+
     }
 }
